gateway/requirements.txt
PyYAML==6.0.2
requests==2.32.3
paho-mqtt==2.1.0
